requirements.txt
email_validator==2.2.0
Flask==3.1.1
Flask-Dance==7.1.0
Flask-Login==0.6.3
Flask-Migrate==4.1.0
Flask-SQLAlchemy==3.1.1
Flask-WTF==1.2.2
python-dotenv==1.1.1
requests==2.32.4
requests-oauthlib==2.0.0
SQLAlchemy==2.0.41
typing_extensions==4.14.1
urllib3==2.5.0
URLObject==2.4.3
Werkzeug==3.1.3
WTForms==3.2.1